// forza-horizon-community.front/src/components/forms/AddNewTuneFormComponent/types.ts
export interface IFormikNewTuneValues {
  title: string;
  forzaShareCode: string;
  selectedCarId: string;
  engineDescription: string;
  engine: string;
  aspiration: string;
  intake: string;
  ignition: string;
  displacement: string;
  exhaust: string;
  handlingDescription: string;
  brakes: string;
  suspension: string;
  antiRollBars: string;
  rollCage: string;
  drivetrainDescription: string;
  clutch: string;
  transmission: string;
  differential: string;
  tiresDescription: string;
  compound: string;
  frontTireWidth: string;
  rearTireWidth: string;
  frontTrackWidth: string;
  rearTrackWidth: string;
}